Hotfixes for the Focus module
Summary:
This differential is a hotfix for the recent work in the Focus module.
One additional class, FITSSkyObject, was extracted and put in its own source files.
The JMIndex parameter of the Centroid detection was not provided to the source extraction procedure.
The Star Profile Viewer missed a few consts in its management of FITS data buffer, which caused some Jenkins builds to fail on SUSE and FreeBSD.

Test Plan: Run kstars_ui_tests

Reviewers: #kstars, mutlaqja

Reviewed By: #kstars, mutlaqja

Subscribers: kde-edu, mutlaqja, murveit

Tags: #kde_edu

Differential Revision: https://phabricator.kde.org/D28689