SEP Focus improvements
Summary:
This PR should improve the HFR calculations with SEP for the focus module and for displaying HFR for those that want to track it in subs.
It is not the end-all, but should be an improvement. There is still work to do.
- Updated and added to tests for the fitssepdetector.cpp
- Removed saturated stars from the HFR calculation, if possible.
- Sped up the SEP star detector 2X when used with full exposures.
- Improved SEP star detector quality, especially for focus/hfr, by (1) changing the deblend constant, and (2) removing the largest stars
- Fixed two bugs in the SEP star detector (1) major issue, ushorts being treated as shorts, and (2) a memory leak for (rare) float images
- Added logging for HFR when calculated
- Fixed bug so that Mark Stars now works
- Improved drawing of circles for Mark Stars, now uses star radius, instead of flakey star width.
- Updated the fitsviewer status to include the HFR when calculated
- Increased a threshold in SEP extract.c which previously would fail on images, e.g. like M42, where there was a lot of area over background.

Not done, but related:
- Should improve the MarkStars UI so that HFR could be calculated without star annotation on display, and could turn on on first image.
- API to star detection methods should be generalized (should wait on rlancaste's investigations of SEP best practices)
- Some display of HFR trends.

Test Plan:
Basic test added to testfitsdata for SEP, run that.
Load a .fits file into fitsviewer. Select view->MarkStars, for a full-exposure sub, you should see ~100 stars circled with an HFR displayed in the status bar.
Load another image and the HFR should continue to be displayed.
Similarly, one could make a capture sequence job with, say 20 5s simulator exposures. After the first exposure completes and the fitsviewer displays
the image, go to the view menu and turn on Mark Stars, then the rest of the images should display with stars circled and HFR displayed in the status bar.
This way of turning on the star/HFR detection should be improved in a future PR.

Reviewers: mutlaqja, lancaster, TallFurryMan

Reviewed By: mutlaqja, lancaster, TallFurryMan

Subscribers: kde-edu

Tags: #kde_edu

Differential Revision: https://phabricator.kde.org/D28767