Fix stretch control bug created in D28935
Summary: When using the stretch controls, bug had image 4X smaller than it should be.

Test Plan:
With D28935, when using the custom stretch sliders, the image displayed while using the sliders
was 4X smaller than it should be (just the preview images--once the sliders were released, the image was
rendered fine). To test, load an image, make a custom stretch change (either changing shadows, midtones or highlights)
and the image displayed while changing the sliders should not change in size as it would without this fix.

Reviewers: mutlaqja

Reviewed By: mutlaqja

Subscribers: kde-edu

Tags: #kde_edu

Differential Revision: https://phabricator.kde.org/D29085