Improve Overlay Rendering
Summary:
Previous PR to reduce fitsview zoom memory use created pixelated rendering for some overlays
with large amount of zoom--e.g. when using a small view (e.g. around a focus view with subframe enabled.
Here we employ a different rendering strategy when the view has a smaller number of pixels (and thus
memory use shouldn't be as much of an issue) so that the rendering is smoother.

Test Plan:
Load fits images, try all the overlays, zoom in and out.
Overlays should not change with images greater than a megapixel.
Try this with various views with subframing (e.g. focus), and the rendering should be less pixelated.

Reviewers: mutlaqja

Reviewed By: mutlaqja

Subscribers: kde-edu

Tags: #kde_edu

Differential Revision: https://phabricator.kde.org/D29192