Fix bug where tab fitsviews weren't receiving images when fitsviewer was not enabled
Summary:
The bug is described in https://indilib.org/forum/ekos/6756-re-problems-with-focus-module-images-not-coming-in.html?start=36#52722
though I actually don't think the particular technique to trigger autofocus matters.
What matters is that fitsviewer is not enabled in the KSTARS FITS settings (not mentioned in that particular thread post).
There was a test for not displaying if fitsviewer was not enabled, but it should have added "but only for NORMAL or CALIBRATION images"
as the other kinds of images have their own private fitsview. There is further complication with batch mode.

Test Plan:
Repeat procedure described in the above forum thread and importantly also disable fitsviewer in the KSTARS FITS options.
Without this fix, the manual focus procedure works, but the automated (after 1 minute) autofocus should get the same
number of stars and HFR calculation for each input image. With this fix, the #stars and HFRs should vary as normal.

Reviewers: mutlaqja

Subscribers: kde-edu

Tags: #kde_edu

Differential Revision: https://phabricator.kde.org/D29218