Fix issue with Meridian Flip Retry
Summary:
The retry timing if fixed so that it waits at least 4 minutes
Mount waits if Capture is in a state where it won't accept the meridian flip request.

Test Plan:
In a situation where the mount doesn't flip (e.g. it needs a larger HA angle before going to the other side of the pier)
this will wait at least the full 4 minutes.
Also, if the capture module is still busy with tasks from the previous failed flip, this should wait until capture is ready to rety.

Reviewers: chrisrowland, mutlaqja

Reviewed By: mutlaqja

Subscribers: kde-edu

Tags: #kde_edu

Differential Revision: https://phabricator.kde.org/D29355