Let UI tests customize the UI settings file.
Summary:
This differential fixes the regression on the user-customized settings file "kstarsui.rc", which was wrongly replaced with the default UI settings file from the resources.
This was done to display UI icons and menus properly in "kstars_ui_tests", but the regression is that any user-customized settings, stored in a local "kstarsui.rc" are not considered anymore.

Test Plan:
Execute "kstars" and check user settings such as keyboard shortcuts are properly enabled.
Execute "kstars_ui_tests" and check icons and menus are properly displayed in the user interface.

Reviewers: #kstars, mutlaqja, murveit

Reviewed By: #kstars, mutlaqja, murveit

Subscribers: kde-edu

Tags: #kde_edu

Differential Revision: https://phabricator.kde.org/D29389