Add an internal guider calibration plot
Summary: It plots the calibration samples, and displays a calibration status message.

Test Plan:
New functionality. Run the internal guider with the simulator.
You can vary the number of iterations used in calibration via the guider parameters.
Click on the new calibration plot (just next to the drift plot)
and you should see updated sample points as calibration points are sampled.
Note, there is a delay at the start when the guide star is being selected.

Reviewers: mutlaqja

Subscribers: kde-edu

Tags: #kde_edu

Differential Revision: https://phabricator.kde.org/D29695