Keep focus temperature until next autofocus is complete
Summary: when starting the autofocus operation, we should not reset
the last focus temperature. This is to avoid the situation of having
autofocus fail or be aborted, and stay long with no temperature
reference for the autofocus on temperature delta feature.

An alternative would have been to just warn the user that we have no
successful autofocus at sequence startup. This was discussed and
discarded in the following merge request:
https://invent.kde.org/education/kstars/-/merge_requests/7

Test plan:
- connect the focuser simulator
- check the temperature delta box
- perform autofocus
- perform a second autofocus and abort it
- create and start a sequence
- increase the temperature in the focuser simulator
- verify that the autofocus is triggered