[Needs testing] Major refactoring of refraction corrections that fixes some bugs
* Potentially introduces some new bugs and therefore needs extensive testing!! *

Potential bugs in SkyMapLite and Android version!

Summary of changes:

SkyPoint does not have a way of differentiating between refracted (apparent) and
actual altitude. The convention has been that the altitude stored in SkyPoint is
always the actual altitude (without refraction corrections) and the refracted
altitude is calculated whenever it is needed.

The reason this matters is because we sometimes call
SkyPoint::HorizontalToEquatorial, which assumes that the altitude in the
SkyPoint is the actual altitude.

However, SkyMap::focus() actually broke this convention! We saved haphazardly
either the refraction-corrected or un-refracted altitude in SkyPoint.

This patch changes the behavior so that SkyMap::focus() is always interpreted as
unrefracted. SkyMap::setFocusAltAz() similarly assumes that the inputs are
actual and not apparent altitudes.

Things I have tested:
1. Slewing of the sky map by mouse
2. Slewing of the sky map by keys
3. Popup menu and some operations therein that use coordinates (Details, DSS...)
4. Enabling and disabling refraction corrections
5. DBus "setRaDec", "setRaDecJ2000" and "setAltAz" methods

Things I have not tested:

1. SkyMapLite / Android versions [although I've tried to port the changes to
   SkyMapLite, there may be things I've overlooked]

2. Equirectangular projection (which is handled rather differently)

3. Ekos / INDI performance

Overall, I believe the change should actually resolve bugs that we didn't know
existed, because of the HorizontalToEquatorial call on refracted coordinates.

I've also added new convenience methods to SkyPoint (SkyPoint::setAltRefracted)
and a bool argument to SkyPoint::refract / SkyPoint::unrefract that helps
simplify code in many places.

Known problems that still exist:
1. "Set Coordinates Manually" does not work as expected
2. The Alt/Az grid is slightly offset from where it should be (Why???)
3. There are computational issues / instability in rendering the Alt/Az grid at the zenith

Potential new problem? Probably not:
Sometimes, there are inconsistencies in the way DeepStars are handled. I get errors like:

WARNING: Marking block with faint mag =  8.71  after block with faint mag  13.56 in trixel 378
Trixel  0 : ERROR: faintMag of prev block =  9.31 , brightMag of block # 1  =  9.92

But I've been unable to reproduce this "before" the patch, but I haven't tried extensively.

CCMAIL: kstars-devel@kde.org