Fix behavior of focus dialog when entering RA/Dec in JNow mode
A problem still exists with the Alt/Az pane. I suspect this is because
of a UI issue -- the tab that we are on is not recorded, and instead
we check if the RA/Dec values in the edit box are okay. Unfortunately,
they are initialized to valid values. This might be the source of the
issue.

Behavior of entering RA/Dec in J2000.0 was already correct.