Autofocus on observatory temperature change
Summary: when the focuser temperature is not available, the autofocus on
temperature change should make use of the observatory temperature as a fallback
(as stated in the v3.4.3 release notes). This is currently not the case.
This commit refactors the focuser and observatory temperatures to make both of
them available to the autofocus trigger feature.

Test plan:
- Connect the observatory simulator
- We need to connect a focuser simulator without temperature handling. To do so,
  I had to comment the temperature in the INDI simulator and recompile it.
- Perform autofocus
- Create and start a sequence
- Change the temperature in the observatory simulator and click the refresh button
- Verify the autofocus is triggered