Add concept of Proxy Devices which are devices that implmenet a different interface of a concrete device decorator.
All INDI devices are initially created as Generic Devices and then decorated to add additional functionality (e.g. Telescope). The telescope driver can also offer additional functionalities that are usually implmeneted in a different interface (e.g. Focuser or Weather). In this implementation, we create a proxy device if we detect that the parent device is not a generic device, but already one that was assigned a specific interface. The difference is that a proxy device does not delete the parent decorator.