Minimize the mean angle representation error of DMS/HMS formats
The methods of class DMS that convert an angle, internally
stored as double, to a DMS or HMS string for display, make a
computation that set the change of the least significant digit
(LSD) aligned with angle integer units. This means that the
rounding error is asymmetric and ranges from zero to 1 arcsec
for DMS or from zero to 1 second for HMS.

The maximum rounding error is made symmetric and halved by
shifting up the angle value by the half value of the LSD.

The high precision DMS and HMS conversions are not changed.