This is a pretty big change.  It is the start of the integration of the StellarSolver Library that I created this spring.   Here are some of the changes that it includes:
Utilizes StellarSolver for both SEP and for Plate Solving.
Adds an Options Profile Editor for creating profiles for use with StellarSolver
Consolidates all the multitudes of options for plate solving and SEP in the Options Profiles.   These profiles are Parameters files that are shared between KStars and StellarSolver and can be used for both SEP and Plate Solving.
Usage of Stellarsolver allows users of all operating systems to plate solve without astrometry.net or an internet connection.  But you do need index files.
Usage of Stellarsolver allows Windows users to now use the index file downloader, directly plate solve using astrometry.net in either Cygwin or ANSVR, and have full control over plate solving.  ANSVR does not need to be "running" and the "online" option is no longer required for using ANSVR.  You can still use that if you like however.
All users can still make use of ASTAP, online Astrometry.net, local astrometry.net, sextractor, and other plate solving features.  I have added more options as well, such as using the internal sextractor with external ASTAP.  These all now share the same options profiles.
Usage of StellarSolver eliminates the need for any external configuration files for Sextractor or Astrometry.net.  You can use internal or external ones without the need for configuration files.
Using the parallel processor options in StellarSolver will allow blind plate solving for many images in just a few seconds instead of minutes.
Several structures are now shared between StellarSolver and KStars so that information can be sent back and forth between them.  These include the Options Profiles (Parameters), Statistic, WCS_Point, Background, and Star (for found sources).
The information contained within the above structures can be retained to add all sorts of new features to KStars for use in Photometry, image analysis, object analysis etc.  In the past, this information was lost as soon as the plate solve or SEP was done.
On the recommendation of Hy, dataType, which was formerly a property of the FITS Data object, is now incorporated into Statistic and I made changes throughout KStars to support this.
The Astrometry Parsers, the Options Maps, some of the settings, and some of the options editor windows have been eliminated because they were redundant, but a lot of the code is still there commented out in case some is still needed.  This needs more work.
For now, the "Remote" option was eliminated because I can't support that with StellarSolver because it depends on INDI to function and that would mean making StellarSolver INDI dependent.  The Remote option could be re-enabled if needed but I think that the fact that KStars can now plate solve on its own without the need for any astrometry.net installation makes it unnecessary.
I have not yet implemented StellarSolver for SEP in Guiding.  I think it will help there too, but I want to make sure that everything else works and check with Hy before I touch his great work there.  He already helped with Focus, but I haven't consulted him yet on Guiding, so it is not implemented there yet.
Load and Slew now loads the image it is trying to solve in the AlignView, and then blind solves with StellarSolver using the options that were selected.  Note that it ignores information in the file such as position or scale.  Load and slew is capable of solving many types of images
"Open FITS Image" in the file menu has been replaced with "Open Image" and is capable of loading any type of image QT can open as well as the FITS types it could before.  This was possible because of the function I had to add to get Load and Slew to work with StellarSolver.

What you can do:

Please check to make sure that this massive update did not break anything important.
Specifically, please verify that none of the code that I commented out or removed when I put StellarSolver in Align or SEP was actually needed after this update.  There was A LOT of code that I had to change and I could have easily messed something up.
Please test StellarSolver to make sure that it plate solves fine in the Align module, and that it works well in the Focus module for SEP in real life situations.  If there are any issues with it, please let me know.
Please help me create good Options Profiles.  StellarSolver comes with a default set of profiles, so there is no initial installation required.  If you change any of them or add any of your own, then it saves a file with the profiles in it. I believe that my profiles for plate solving are pretty good, but that doesn't mean they can't be improved upon.  I think that my profiles meant for SEP need some work.  I haven't gotten a lot of feedback on them so far.  You can easily change them and you can create your own options profiles.  I would like to provide a good default set for users.  If you find anything in the default set that you think would be better, or if you come up with a profile that you think works better for certain situations, I would like to include that, please send it to me.

Testing instructions:

First, please download and install StellarSolver.  Use the latest Master because I have been making some changes recently based on things that I encountered while integrating into KStars and I am still updating based on that.  Those changes are not in the latest release yet.  https://github.com/rlancaste/stellarsolver. You can use the install script if you like or you can build and install it manually.  For Windows and Mac testing, the repository includes a craft recipe for installing in Craft on Windows and Mac OS.
Then you can test the code in my merge request.  One option to test it would be to download the code in my fork and then build and install KStars normally.  Another would be to use the features in GitLab to download a patch and then apply it to a local copy of the latest KStars you have.  Since this is my first merge request since we moved to Gitlab, I don't know all the details of how to do the second option.