Fix HFR check breaking running autofocus
This MR prevents the autofocus procedure from breaking when Capture
resumes after requiring adjustment from Focus before guiding deviation
became unacceptable. In that situation Capture was asking for an HFR
check, which was leading to a Focus capture, which in turn was breaking
the autofocus procedure. This required manual intervention to recover.

This MR is pushed in the context of #53, and only prevents Focus from
hanging. It does not prevent Capture from requiring a refocus every time
guiding deviation returns to normal if it is needed, which is an
acceptable behaviour.