Test Scheduler::processJobInfo in preparation of issue #54
The private method Scheduler::processJobInfo was copied as a public
member of TestPlaceholderPath::testSchedulerProcessJobInfo.
However, I replaced the second parameters SchedulerJob by just the
targetName of the SchedulerJob (could not manage to instantiate a
SchedulerJob).

I have identified the variables that modifies the signature of a
SequenceJob and added the test data in testSchedulerProcessJobInfo_data.csv
since the test data is quite big.
This data file is then added to the executable using the qt ressource system
and later read at runtime in TestPlaceholderPath::testSchedulerProcessJobInfo_data.