A New Dark Frame Library
This MR introduces a complete revamp of the dark frame library. A new wizard is provided in the capture module that enables the user to capture traditional dark frames using a number of exposures, binning, and temperatures in order create one master dark frame for these settings.

Furthermore, for non-cooled camera, defect maps offer better denoisng potential where the traditional dark subtration method struggle to remove the hot and cold pixels present in the frame. The user can examine and save defect maps for various dark frames and then select which method should be used for each camera.

Significant reductions in noise were observed in sample tests which led to improved reliability of the various star detection algorithms in KStars.