Stable/Unstable tests.
This MR divides tests into stable, which are mandatory for any MR to be merged, and unstable, which are still in development and are allowed to fail. It is of course desirable to have a maximum of tests in the stable category.

One problem preventing parallel execution of stable/unstable tests is that `ctest` does not allow output to a custom XML file. The output goes to `Testing/YYYYMMDD-HHMM/Test.xml`, and the minute granularity is not very helpful. So stable tests are done first in stage `test` and unstable are done in a further stage, `test-unstable`. Thus, if on the expedite line, a MR may be merged as soon as the `test` stage is done.