.gitlab-ci.yml: force build-and-test-unstable to succeed
My queue-size script estimates how much work is pending in the GitLab CI
queue based on the runtime of past *successful* jobs with the same name.
The build-and-test-unstable job never ran successfully, so I never get an
estimated ETA for it.

This commit forces it to succeed with "|| true" to get a single
"successful" job, which will be used in future estimates. This is not
intended to be merged to master or even go in a merge request, it's just
to make gitlab run the jobs once.