Remove catalog color for OpenNGC. Apply colorscheme-specified colors consistently in the absence of a catalog color.
Finding a color that works across all three of Night Mode / Dark
background / Light background is hard. So the proposal here is to
remove the default color for OpenNGC, so that our default catalog
works correctly in all color schemes.

The remainder of the changes are needed to consistently apply the
color scheme when dynamically switching color schemes in the app (as
opposed to picking up the color at the start and sticking with it)