Fixing Data directory install to app bundle for macs
A user detected an issue where the planets were missing from the Mac version of kstars.  I realized that the reason for this was because by just copying the data directory into the app bundle instead of "installing" it, there were a couple of files that were not put in the right sub directory as a result and there were several extra files included that should not be there.  This fix should make kstars actually install to the directory we want to put the data directory in, instead of just copying the source directory.  Please verify that it still works fine on Linux and Windows.