dither and calibration cleanup.
The dithering process calculates a random offset in pixels and moves that amount using the same mechanism as guide corrections. Since this is a random movement, exactly meeting the random dither amount isn't critical, and takes time and can potentially stall the system. Therefore, this MR removes the guide/dither iterations that check if the mount is pointing very closely to the desired dither location. It simply calculates the pulses it expects to need to move the mount the desired random amount and issues those pulses without checking that the movement was done. It records the observed position, of course, as the target guide position. This is enabled with a new checkbox option in the guider/dither options menu, but I'm tempted to remove the checkbox from this MR and just make the behavior the default. 

Another change is that previously, if one of the guide axes or directions were disabled, the dithering would not pulse in that direction (which can cause a nasty dithering issue currently, see https://indilib.org/forum/ekos/11008-dither-failure-while-guiding-is-good.html#78956
This MR changes it so that both RA and DEC are always dithered, even if an axis is disabled.
We could instead make this an option, or even add separate enable/disable-axis buttons for dithering. That can be up for discussion.

Finally, when testing the new dithering behavior using the simulator, I found that the dithering movements made weren't as close to the desired random dither as I'd expected. I traced this to the fact that the simulator has non-square guider-camera pixels. There was a TODO in the calibration code to improve matters with non-square pixels, and it was not difficult to implement, so I corrected this as well. Now the new dithering with the simulator operates well, as I expected.


- I have tested this with the simulator, but not with my telescopes (due to weather). I can/will test in a few weeks, or feel free to test yourself and let me know how it goes.

- Please feed back whether you're comfortable that always dithering in both axes is OK, even with the axis disabled for guiding. Alternatively whether I should add enable/disable axis checkboxes to the guider/dither options menu.