Don't allow guidestars to redo star correspondence if it isn't the first guide frame
This is an attempt to fix the bug described in 
https://indilib.org/forum/ekos/11029-internal-guider-shift-when-loosing-guide-start-with-cloud.html?start=12

What I think is happening is that GuideStars::findGuideStar() was allowed to re-define the guide star if there were no multi-star StarCorrespondence setup, because of in the logic where the scheduler was restarting the guider, selectGuideStar() was not being called. This normally works OK. However, in poor skies it, if not enough stars detected (there needs to be a minimum of 5) then starCorrespondence is not established, and the selectGuideStar can find a new/different guide star. Even this would be OK if the guider realized that this was a new guide star, but it doesn't know that and tries to move the two stars to the same image position.

The solution here is to not allow findGuideStar to redefine the guide star unless it's the first guider iteration (InternalGuider::m_isFirstFrame is true).

A possible weakness is that if the guider is started in poor skies MultiStar won't be used, but then if the skies clear up and MultiStar is now possible, it would not switch over to MultiStar. I suppose something like that could be engineered.

**Why Draft**

This is marked as Draft as there has been minimal testing, just quick simulator--I have not been able to test on a telescope. Am creating the MR in hopes of getting other to test until my skies clear up.