Deactivating meridian flip while PAA is running
Running a polar alignment close to the meridian might lead to the situation that the scope crosses the meridian. In this case, the meridian flip feature of `Mount` will be deactivated so that the polar alignment is not disturbed.

As soon as the polar alignment is finished, the meridian flip feature is activated again.