Add support for flat darks.
Add support to dark flats. Steps:

1. Add flats
2. When done, click on "Generate Flat Darks"
3. Dark flat frames are generated and appended.
4. Start sequence. 

Dark flat frames would be captured for the same duration of the flat frames. The only limitation to this method is that it need to be performed in the same session. i.e. you cannot take some flats, restart Ekos, and then continue since we can no longer know the duration that was calculated for the flat frames. While it is possible to open the FITS files and inspect them, this is very complicated now and perhaps a better solution for such case is available in the future (maybe a file to track the flat frame exposures).