Bugfix: test to delay capture with high guide deviation was broken
The test to (optionally) not start a capture is the guide deviation is greater than a certain amount was broken for a while. It never triggered. This MR fixes that. 

The issue was that the test, internal to sequenceJob/sequenceJobState, checks to see if guiding is active before applying the test. However, when capture sees guider change state to active (in Capture::setGuideStatus) it cannot change the guide state internal to the sequence job, because currentJob is still a nullptr, not yet set to the current job. This change fixes that by always setting the sequenceJob's guide state before emiting prepareCapture().