Adjust dbus interfaces and paths to allow test_ekos_scheduler_ops tests to...
Adjust dbus interfaces and paths to allow test_ekos_scheduler_ops tests to succeed when Ekos is running.

Made a new dbus, called org.kde.mockkstars with base path /MockKStars
This prevents dbus collisions that would occur when a standard Ekos dbus connection is running on the same computer--e.g. an instance of Ekos is running and at the same time the test_ekos_scheduler_ops test tries to run.

**TESTING THIS CHANGE**

**Before merging in this new code**

- End all running KStars/Ekos jobs on your computer.
- Then run test_ekos_scheduler_ops. The test should pass.
- Now start KStars, Start Ekos, connect to Indi.
- Then run test_ekos_scheduler_ops again. The test should fail.

**After merging in this new code**
- Same as above, except the test_ekos_scheduler_ops test should pass on the last line.