Analyze: update graphics plot for completed capture or focus if latest is checked.
Analyze has a checkbox labeled "Latest" that applies to when Analyze is monitoring the current Ekos session. Its function is to keep the most recent data (e.g. the recent captures, focus sessions and guide deviations) always displayed on the UI. With this MR, when "Latest" is checked and a capture completes, the capture's guiding data is automatically plotted. Similarly, when a focus completes, its focus curve is plotted.