guider cleanup and bugfix
Main change fixes a bug where the guider, if having trouble detecting stars when reacquiring the guide star, or starting up, can wind up with a reticle of 0,0 and push the image towards the edge. 

Also cleans up other parts of the guider.

Tested successfully tonight, including putting a hat on the guidescope several times for several minutes and can no longer get the guider to push the image away from the target as it had been doing in poor conditions.