Check position after captures
This is early work on regularly checking the position of a capture to make sure it hasn't drifted (for various reasons) too far from the target position.

At this point there is no UI to set parameters and no action taken if the drift is large. Simply after each LIGHT capture the position is solved and compared with the target position, and printed into the log (and the space at the bottom of the capture tab).

The system limits the time the solver can take, and automatically makes sure it is less than 15s and 2s less than the exposure time, and runs on each finished LIGHT capture.

I found that (at least on the simulator) the target signal wasn't set properly by the mount module, so for this to be accurate, you'd need to make sure that there is an Alignment before the capture sequence, which is done most of the time anyway. Of course, this can be fixed in mount if it is a problem.

Look for this log line:

```
2022-02-02T23:16:22 Position check took 0.51s. Distance to target: 18 arcsec. Target 09h 23m 36s  50° 52' 51"  Position 09h 23m 35s  50° 52' 46"  (dRA -17 dDEC -6)
```

I'm a little unsure of the J2000 vs JNow conversions that are done.

My idea is that ultimately, the user would specify a tolerance in arc-seconds or arc-minutes and if that is exceeded (perhaps twice in a row) then capture would call abort() or something like that, and the scheduler would take over and restart the job, hopefully aligning before capture restarts.

This is related to Scott Denning's post: https://indilib.org/forum/wish-list/11217-wishlist-more-failure-please.html#80174

@wreissenberger @mutlaqja in particular should check this out.