Bugfix darks jobstate and #158
There is a bug in the reworked sequence job state machine that does not always set the job state to "busy" when a job preparation is started. This has a side effect on dark library creation where saving the master file does not work.

This is fixed with this change. Additionally, the issue #158 is also fixed.