Unconstrain Ekos' main vertical slider
We have been getting complaints forever about the vertical slider in Ekos. Specifically it is hard/impossible to increase the size of the log window at the bottom. A recent example is https://indilib.org/forum/ekos/11283-increase-height-of-the-scrolling-info-pane-at-the-bottom-of-the-ekos-window.html

The issue is that the various module tabs have minimum vertical sizes, and given a user's overall Ekos window vertical size, there's typically no room left to expand the log window. I believe typical use is to peek into the log to see what's going on, and then to restore to a similar height as was originally.

This change allows the slider to overrule the minimum vertical size requirement for the Ekos tab, and so the user can expand the log window and squash the Ekos tab as much as the user wants. For sure, this will allow the user to squash it so that it is unusably squashed, but it will allow the user to see the log and then to restore to a readable vertical size. I believe this is a better result that the existing inflexible slider.