Make default color scheme default
KStars has a default color scheme, which is named "Default", but it's
not actually the default. Instead, the actual default is "Black Body",
which unfortunately suffers from a significant usability issue: the
menubar text is unreadable with the Breeze theme. This affects users
using the Breeze theme, as well as users using a Flatpak package of the
app.

The path of least resistance is to make the color scheme called
"Default" actually be the default, and this seems to make some logical
sense given its name. So let's do that.

To test, use the Breeze widget theme, remove the value of `CurrentTheme` from your `kstarsrc` file, and re-launch the app.

| Before | After |
| ------ | ----- |
| ![kstars_before](/uploads/6bf9142e483889a9e30e135957aa5fa7/kstars_before.png) | ![kstars_after](/uploads/62a2b9f9f41f90bffaa0b59fd85b146a/kstars_after.png) |